requirements.txt
trimesh
scikit-learn
opencv-python
matplotlib
imageio
diffusers==0.26.3
einops
transformers==4.43.2
open-clip-torch
gradio
pytorch-lightning==1.9.1
omegaconf
triton
accelerate
objaverse
mathutils
iopath==0.1.10
mediapipe
xatlas
rembg
lightning==2.3.3
numpy==1.26.4
wandb
plotly
pyrender
open3d
numpy-quaternion
pygraphviz
openai
python-dotenv